3-13. List 컴포넌트 생성하기
- 할 일 목록의 각 할 일 컴포넌트화
