+++ src/App.js
@@ -1,6 +1,6 @@
 import React, {useState} from 'react';
 import './App.css';
-import List              from './components/List';
+import Lists             from './components/Lists';
 import Form              from './components/Form';
 
 export default function App() {
@@ -29,7 +29,7 @@ export default function App() {
           <h1>할 일 목록</h1>
         </div>
 
-        <List todoData={todoData} setTodoData={setTodoData}/>
+        <Lists todoData={todoData} setTodoData={setTodoData}/>
 
         <Form handleSubmit={handleSubmit} value={value} setValue={setValue}/>
       </div>
